Add options for easily booting from a new volume
Previously, the only option to boot from a volume using OpenStack
client consisted of using the --block-device-mapping option. This
option requires that the volume already exists prior running the
server create command as per the v1 specification of block device
mapping. Since the addition of the v2 block device mapping format
the compute API can create the volume from a source type such as
an image or a snapshot, and attach it to use it as the boot disk
for the new instance, all during the create request.

Instead of simply exposing the block device mapping v2 format
directly, this exposes a set of options that allows the user to
specify a boot-disk-type which maps to the destination type in
the v2 format, and it the source type is assume to always be either
an image or snapshot. For more complex block device mapping, then
it's suggested to use the --block-device-mapping option.

Exposing the block device mapping v2 format direclty
can be cumbersome to use since permitted values cannot be easily
discovered and does not indicate which are required and which are
optional without out of band knowledge (docs).